+++ packages/devtools_app/lib/src/screens/memory/panes/diff/diff_pane.dart
@@ -149,7 +149,8 @@ Find unexpected memory usage by comparing two heap snapshots:
 
     b. Execute the feature in your application
 
-    c. Take a second snapshot
+    c. Take a second snapshot. If you are experiencing DevTools crashes due to size of snapshots,
+       switch to the [desktop version](https://github.com/flutter/devtools/blob/master/BETA_TESTING.md).
 
     d. While viewing the second snapshot, click **Diff with:** and select the first snapshot from the drop-down menu;
     the results area will display the diff