+++ tool/release_helper.sh
@@ -47,16 +47,17 @@ git commit -a -m "$COMMIT_MESSAGE"
 
 git push -u $DEVTOOLS_REMOTE $RELEASE_BRANCH
 
-echo "Creating the PR"
+echo "$0: Creating the PR"
 echo
 
 PR_URL=$(gh pr create --repo flutter/devtools --draft --title "$COMMIT_MESSAGE" --fill)
 
-
-echo "Updating your flutter version to the most recent candidate."
+echo "$0: Updating your flutter version to the most recent candidate."
 echo
 ./tool/update_flutter_sdk.sh --local
 
-echo "Your Draft release PR can be found at: $PR_URL"
+echo "$0: Your Draft release PR can be found at: $PR_URL"
 echo
+echo "$0: DONE.
 echo "$0 DONE: Build, run and test this release using: 'dart ./tool/build_e2e.dart'"
+