+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.3-wip
 * Add `showNotification` and `showBannerMessage` endpoints to the extensions API.
 * Add hot reload and hot restart actions to the simulated DevTools environment.
+* Update `build_and_copy` command, as well as documentation, to reference `config.yaml`
+instead of `config.json`, as required by `package:extension_discovery` v2.0.0.
+* Add exception handling to `devtools_extensions build_and_copy` command.
 
 ## 0.0.2
 * Add a simulated DevTools environment that for easier development.